Add files via upload
The goal of the following algorithm is to identify breast cancer, using only a small fraction of the available data (N=1148 for training, N=163 for testing and N=382 for the validation set) from the BreakHis dataset. To do so, we will use a Transfer Learning technique using the Xception ConvNet.